Improve the error message when a column is missing from the schema
